+++ packages/sit-onyx/src/index.ts
@@ -62,8 +62,8 @@ export * from "./components/OnyxMiniSearch/types";
 export { default as OnyxMobileNavButton } from "./components/OnyxMobileNavButton/OnyxMobileNavButton.vue";
 export * from "./components/OnyxMobileNavButton/types";
 
-export { default as OnyxMore } from "./components/OnyxMore/OnyxMore.vue";
-export * from "./components/OnyxMore/types";
+export { default as OnyxMoreList } from "./components/OnyxMoreList/OnyxMoreList.vue";
+export * from "./components/OnyxMoreList/types";
 
 export { default as OnyxNavAppArea } from "./components/OnyxNavAppArea/OnyxNavAppArea.vue";
 export * from "./components/OnyxNavAppArea/types";
@@ -124,7 +124,6 @@ export * from "./components/OnyxVisuallyHidden/types";
 
 export { default as OnyxForm } from "./components/OnyxForm/OnyxForm.vue";
 export * from "./components/OnyxForm/types";
-
 export * from "./composables/density";
 export * from "./composables/scrollEnd";
